feat: change min go version to 1.18
