+++ .github/workflows/build.yml
@@ -19,7 +19,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        go-version: [1.17, oldstable, stable]
+        go-version: [1.18, oldstable, stable]
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
