+++ env.go
@@ -191,6 +191,28 @@ func ParseWithOptions(v interface{}, opts Options) error {
 	return parseInternal(v, setField, customOptions(opts))
 }
 
+// ParseAs parses the given struct type containing `env` tags and loads its
+// values from environment variables.
+func ParseAs[T any]() (T, error) {
+	var t T
+	return t, Parse(&t)
+}
+
+// ParseWithOptions parses the given struct type containing `env` tags and
+// loads its values from environment variables.
+func ParseAsWithOptions[T any](opts Options) (T, error) {
+	var t T
+	return t, ParseWithOptions(&t, opts)
+}
+
+// Must panic is if err is not nil, and returns t otherwise.
+func Must[T any](t T, err error) T {
+	if err != nil {
+		panic(err)
+	}
+	return t
+}
+
 // GetFieldParams parses a struct containing `env` tags and returns information about
 // tags it found.
 func GetFieldParams(v interface{}) ([]FieldParams, error) {
