+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/caarlos0/env/v10
 
-go 1.17
+go 1.18
